Implemented tag share subcommand

+++ modules/Misc/tag/remove.js
@@ -11,8 +11,13 @@ module.exports = async function(msg, args) {
     if (rows.length < 1)
         throw 'I can\'t delete a tag that doesn\'t exist!';
 
-    // Delete row from db and if image exists in tagImages, delete it too.
+    // Delete row from db
     await db.deleteRows('tags', rows[0]);
-    if (rows[0].imageid)
-        await tagUtils.deleteImage(rows[0].imageid);
+
+    // Delete the image if it exists
+    // and there are no other users with the same tag
+    if (rows[0].imageid) {
+        if(!(await db.ifRowExists('tags', { 'imageid': rows[0].imageid })))
+            await tagUtils.deleteImage(rows[0].imageid);
+    }
 }
